Start changing the Pokemon data system to be cleaner, and use singletons
The main things in this update are changing the main parts of Pokemon data to be in a single json file instead of multiple, changing the entire system for how Pokemon data is stored and other data to use enums instead.
I also started to change different classes that are widely used to Singletons instead to make things cleaner.